refactor: extract validationManager to another file
